feat: cleanup and pass correct event ids
